Move App and Root to app/containers
