+++ app/containers/Root.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Provider } from 'react-redux'
-import App from 'app/containers/App'
+import App from 'app/App'
 
 const Root = ({ store }) => (
   <Provider store={store}>
