+++ app/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from 'react-dom'
 import createStore from 'app/redux'
-import Root from 'app/Root'
+import Root from 'app/containers/Root'
 
 const store = createStore()
